builder: expose build parameters 'transform' and 'exclude'
Using these allows for more elaboarated builds, and excluding 'dgr'
from being part of the ACI image. For example, this is how you do it:

  build:
    exclude:
      - /dgr

To create a replacement tarball for the core 'aci-builder' you'd:

  build:
    exclude:
      - /etc/resolv.conf
      - /dgr
    transform:
      - 's@^rootfs@rootfs/dgr@'

If a path starts with a slash then the correct 'rootfs' or its alias
gets prepended automatically.

Please keep in mind that 'tar' applies exclusions first, followed by
transformations.