builder: improve legibility of JSON manifest files
A JSON manifest file often contains many lines with {name…, value…}
pairs, which are tedious to read. Henceforth they will be arranged
on one line per pair.

HTML-escaping of some chars, most notably < and > in email addresses,
has been disabled. Manifests are unlikely to be read by browsers, but
some languages are troublesome with converting those values back.
This change is currently without effect because the YAML to JSON
conversion library which we use triggers the escaping still.